src/test/java/com/hotelreservation/HotelReservationTest.java
package com.hotelreservation;

import static org.junit.Assert.assertEquals;

import java.text.ParseException;
import java.util.List;

import org.junit.Test;

public class HotelReservationTest {
    @Test
    public void WhenHostelsAreAddedFindCheapestHotel() throws ParseException {
        Hotel hotel1 = new Hotel("Lakewood", 110, 90);
        Hotel hotel2 = new Hotel("Bridgewood", 160, 60);
        Hotel hotel3 = new Hotel("Ridgewood", 220, 150);

        HotelReservation hotelReservation = new HotelReservation();
        hotelReservation.addHotel(hotel1);
        hotelReservation.addHotel(hotel2);
        hotelReservation.addHotel(hotel3);
        Hotel cheapestHotel = hotelReservation.findCheapestHotel();
        long totalDays = hotelReservation.getTotalNoOfDays("10Sep2020", "11Sep2020");
        long totalCost = cheapestHotel.getWeekDayRateRegCus() * totalDays;
        assertEquals(220, totalCost);
        assertEquals("Lakewood", cheapestHotel.getHotelName());
    }

    @Test
    public void WhenHostelsAreAddedReturnSize() {
        Hotel hotel1 = new Hotel("Lakewood", 110, 90);
        Hotel hotel2 = new Hotel("Bridgewood", 160, 60);
        Hotel hotel3 = new Hotel("Ridgewood", 220, 150);

        HotelReservation hotelReservation = new HotelReservation();
        hotelReservation.addHotel(hotel1);
        hotelReservation.addHotel(hotel2);
        hotelReservation.addHotel(hotel3);
        assertEquals(3, hotelReservation.totalHotels());
    }
    @Test
    public void WhenHostelsAreAddedFindCheapestHotelBasedOnWeekDayAndWeekEndRates() throws ParseException {
        Hotel hotel1 = new Hotel("Lakewood", 110, 90);
        Hotel hotel2 = new Hotel("Bridgewood", 160, 60);
        Hotel hotel3 = new Hotel("Ridgewood", 220, 150);

        HotelReservation hotelReservation = new HotelReservation();
        hotelReservation.addHotel(hotel1);
        hotelReservation.addHotel(hotel2);
        hotelReservation.addHotel(hotel3);
        List<String> cheapestHotelList = hotelReservation.findCheapestHotelBasedOnWeekEndAndWeekDaysOffer("11sep2020",
                "12sep2020");
        String cheapestHotel = "";
        for (String name : cheapestHotelList) {
            cheapestHotel = name;
        }
        assertEquals("Lakewood", cheapestHotel);
    }
}